+++ R/plot_graphr.R
@@ -1,5 +1,3 @@
-# Single graphs ####
-
 #' Easily graph networks with sensible defaults
 #' 
 #' @description 
@@ -129,8 +127,7 @@ graphr <- function(.data, layout, labels = TRUE,
   }
   if (missing(node_group)) node_group <- NULL else {
     node_group <- as.character(substitute(node_group))
-    g <- tidygraph::activate(g, "nodes") %>%
-      tidygraph::mutate(node_group = reduce_categories(g, node_group))
+    g <- manynet::mutate_nodes(g, node_group = reduce_categories(g, node_group))
   }
   if (missing(edge_color) && missing(edge_colour)) {
     edge_color <- NULL
@@ -176,12 +173,11 @@ graphr <- function(.data, layout, labels = TRUE,
   p <- ggraph::ggraph(lo) + ggplot2::theme_void()
   if (!is.null(node_group)) {
     x <- y <- NULL
-    thisRequires("ggforce")
-    thisRequires("concaveman")
+    # thisRequires("ggforce")
     p <- p + 
       ggforce::geom_mark_hull(ggplot2::aes(x, y, fill = node_group,
                                            label = node_group), data = lo) +
-      ggplot2::scale_fill_manual(values = colorsafe_palette,
+      ggplot2::scale_fill_manual(values = ag_qualitative(length(unique(p$data[[node_group]]))),
                                  guide = ggplot2::guide_legend("Group"))
   }
   if(snap){
@@ -222,7 +218,7 @@ graphr <- function(.data, layout, labels = TRUE,
                                                      ifelse(is.null(edge_color) &
                                                               manynet::is_signed(g),
                                                             "Edge Sign", "Edge Color")))
-    } else p <- p + ggraph::scale_edge_colour_manual(values = getOption("snet_cat", default = colorsafe_palette),
+    } else p <- p + ggraph::scale_edge_colour_manual(values = ag_qualitative(length(unique(out[["ecolor"]]))),
                                                    guide = ggplot2::guide_legend(
                                                      ifelse(is.null(edge_color) &
                                                               manynet::is_signed(g),
@@ -249,7 +245,7 @@ graphr <- function(.data, layout, labels = TRUE,
         p <- p + ggplot2::scale_colour_manual(values = getOption("snet_highlight", default = c("grey","black")),
                                               guide = ggplot2::guide_legend("Node Color"))
       } else {
-        p <- p + ggplot2::scale_colour_manual(values = getOption("snet_cat", default = colorsafe_palette),
+        p <- p + ggplot2::scale_colour_manual(values = ag_qualitative(length(unique(out[["ncolor"]]))),
                                               guide = ggplot2::guide_legend("Node Color"))
       }
     }
@@ -276,24 +272,29 @@ graphr <- function(.data, layout, labels = TRUE,
       hj <- ifelse(p[["data"]][,1] >= 0, -0.2, 1.2)
     }
     p <- p + ggraph::geom_node_text(ggplot2::aes(label = name), repel = TRUE,
+                                    family = ag_font(),
                                     size = 3, hjust = hj, angle = angles) +
       ggplot2::coord_cartesian(xlim=c(-1.2,1.2), ylim=c(-1.2,1.2))
   } else if (layout %in% c("bipartite", "railway") | layout == "hierarchy" &
              length(unique(p[["data"]][["y"]])) <= 2) {
     p <- p + ggraph::geom_node_text(ggplot2::aes(label = name), angle = 90,
+                                    family = ag_font(),
                                     size = 3, hjust = "outward", repel = TRUE,
                                     nudge_y = ifelse(p[["data"]][,2] == 1,
                                                      0.05, -0.05)) +
       ggplot2::coord_cartesian(ylim=c(-0.2, 1.2))
   } else if (layout == "hierarchy" & length(unique(p[["data"]][["y"]])) > 2) {
     p <- p + ggraph::geom_node_text(ggplot2::aes(label = name),
+                                    family = ag_font(),
                                     size = 3, hjust = "inward", repel = TRUE)
   } else if (layout %in% c("alluvial", "lineage")) {
     p <- p + ggraph::geom_node_label(ggplot2::aes(label = name), size = 3,
+                                     family = ag_font(),
                                      repel = TRUE, nudge_x = ifelse(p[["data"]][,1] == 1, 
                                                                     0.02, -0.02))
   } else {
     p <- p + ggraph::geom_node_label(ggplot2::aes(label = name),
+                                     family = ag_font(),
                                      repel = TRUE, seed = 1234, size = 3)
   }
 }