+++ tests/testthat/test-theme_match.R
@@ -1,3 +1,11 @@
+test_that("match colors works", {
+  expect_equal(match_color("#4575b4", c("blue","red")), "blue")
+})
+
+test_that("is dark works", {
+  expect_equal(unname(is_dark(c("#000","#FFF"))), c(TRUE, FALSE))
+})
+
 test_that("Single color input returns correct match", {
   res <- match_color("#4575b4", c("#006564", "#0094D8", "#622550"))
   expect_length(res, 1)